Add POC armour scaling.

+++ src/server/game/Entities/Creature/CreatureData.h
@@ -268,9 +268,9 @@ struct TC_GAME_API CreatureBaseStats
         return uint32(ceil(BaseMana * info->ModMana));
     }
 
-    uint32 GenerateArmor(CreatureTemplate const* info) const
+    float GenerateArmor(CreatureTemplate const* info) const
     {
-        return uint32(ceil(BaseArmor * info->ModArmor));
+        return ceil(BaseArmor * info->ModArmor);
     }
 
     float GenerateBaseDamage(CreatureTemplate const* info) const
